backend/requirements.txt
Flask
pymongo
flask-cors
Flask-JWT-Extended
passlib
python-dotenv
